pyproject.toml
[project]
name = "tools-config"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.10"
dependencies = [
    "pypdf>=5.4.0",
    "pyyaml>=6.0.2",
]
